refactor: 헷갈리지 않게 폴더명 등은 Apply, 라우팅은 recruit
- url navigation이 필요한 부분에는 주석 달아놨습니다!
