+++ src/components/Header/Header.tsx
@@ -64,7 +64,12 @@ export default function Header() {
               </S.NavItem>
             </S.Nav>
           </S.NavWrapper>
-          <S.FarmingLogButton onClick={() => setPopupOpen(true)}>지원하기</S.FarmingLogButton>
+          <S.FarmingLogButton 
+            // 지원하기 url은 '/recruit'입니다!
+            onClick={() => setPopupOpen(true)}
+          >
+            지원하기
+          </S.FarmingLogButton>
         </>
       )}
 
@@ -111,6 +116,7 @@ export default function Header() {
               <S.NavItem 
                 $isTablet={isTablet} 
                 $isMobile={isMobile} 
+                // 지원하기 url은 '/recruit'입니다!
                 onClick={() => setPopupOpen(true)} 
                 isActive={false}
               >
